hw/hal: Use 32 bit for flash alignment
This changes alignment type from uint8_t to uint32_t
in several places.

For flash_area_align() function 32 bits are already used
by other platform supported by mcuboot.

struct hal_flash also was using 32 bits before.

+++ hw/hal/include/hal/hal_flash.h
@@ -159,7 +159,7 @@ int hal_flash_isempty_no_buf(uint8_t id, uint32_t address, uint32_t num_bytes);
  *
  * @return                      The flash device's minimum write alignment.
  */
-uint8_t hal_flash_align(uint8_t flash_id);
+uint32_t hal_flash_align(uint8_t flash_id);
 
 /**
  * @brief Determines the value of an erased byte for a particular flash device.
